services/parser/requirements.txt
celery>=5.3
pyyaml>=6.0
requests>=2.32
python-dotenv>=1.0.1
